Semantic Logo Markup
- Reposition the alt attribute. The alt attribute is only for use on
the img element. Since it gives a text-based description of the images
content, it isn’t needed on elements like div, whose content is
supposed to already be text.

- Make the words “Cylinder Project” part of an h1, to signal importance
(it is the page’s top-level heading)

- Remove the wrapper and masthead divs (neither should be necessary)

- Now that the logo comes in the form of one image (rather than two
background images), I changed the source to be the cylinder_3_large.png
file, and deleted the two masthead images. Using the same logo file on
both pages, and having the text and the icon in one file, will speed up
page load time.

- Set the image’s width in css to always be 100% of its container, so
that it resizes appropriately.

- remove the css for the old background images.